Fix feature preview logic for toggling visibility of private channel toggle in realtime settings
