+++ apps/studio/components/layouts/RealtimeLayout/RealtimeLayout.tsx
@@ -4,10 +4,9 @@ import { PropsWithChildren } from 'react'
 import { ProductMenu } from 'components/ui/ProductMenu'
 import { useSelectedProject } from 'hooks/misc/useSelectedProject'
 import { withAuth } from 'hooks/misc/withAuth'
-import { useFlag } from 'hooks/ui/useFlag'
+import { useIsRealtimeSettingsEnabled } from 'hooks/ui/useFlag'
 import ProjectLayout from '../ProjectLayout/ProjectLayout'
 import { generateRealtimeMenu } from './RealtimeMenu.utils'
-import { useRealtimePrivateOnlyModeEnabled } from 'components/interfaces/App/FeaturePreview/FeaturePreviewContext'
 
 export interface RealtimeLayoutProps {
   title: string
@@ -15,19 +14,7 @@ export interface RealtimeLayoutProps {
 
 const RealtimeLayout = ({ title, children }: PropsWithChildren<RealtimeLayoutProps>) => {
   const project = useSelectedProject()
-  const isRealtimePrivateOnlyModeEnabled = useRealtimePrivateOnlyModeEnabled()
-  // the flag is used to enable/disable the realtime settings for all projects. Will be overridden by the isRealtimePrivateOnlyModeEnabled if it is enabled.
-  const enableRealtimeSettingsFlag = useFlag('enableRealtimeSettings')
-  // the flag is used to enable/disable the realtime settings for specific projects. Will be overridden by the enableRealtimeSettingsFlag if it is enabled.
-  const approvedProjects = useFlag<string>('isRealtimeSettingsEnabledOnProjects')
-
-  const isEnabledOnProject =
-    !!project?.ref &&
-    typeof approvedProjects === 'string' &&
-    (approvedProjects ?? '')
-      .split(',')
-      .map((it) => it.trim())
-      .includes(project?.ref)
+  const enableRealtimeSettings = useIsRealtimeSettingsEnabled()
 
   const router = useRouter()
   const page = router.pathname.split('/')[4]
@@ -39,9 +26,7 @@ const RealtimeLayout = ({ title, children }: PropsWithChildren<RealtimeLayoutPro
       productMenu={
         <ProductMenu
           page={page}
-          menu={generateRealtimeMenu(project!, {
-            enableRealtimeSettings: isRealtimePrivateOnlyModeEnabled || enableRealtimeSettingsFlag || isEnabledOnProject,
-          })}
+          menu={generateRealtimeMenu(project!, { enableRealtimeSettings })}
         />
       }
     >
